+++ src/commands.ts
@@ -90,6 +90,15 @@ export function registerCommands(
                     });
 
                     await WEB_VIEW.open();
+                } else if (RESULT && _.isSymbol(RESULT['__neweditor_tm_19790905'])) {
+                    const DOC = await vscode.workspace.openTextDocument({
+                        content: RESULT.text,
+                        language: RESULT.lang,
+                    });
+
+                    await vscode.window.showTextDocument(
+                        DOC, RESULT.column,
+                    );
                 } else if (Buffer.isBuffer(RESULT)) {
                     let md = '# Code Execution Result (Buffer)\n\n';
                     md += '```';