+++ src/tools/quickcode.ts
@@ -483,7 +483,64 @@ ${ $h.toStringSafe(DOCUMENT.getText()) }
             return {
                 '__markdown_tm_19790905': Symbol('MARKDOWN_DOCUMENT'),
                 'markdown': markdown.toString('utf8'),
+                'toString': function() {
+                    return this.markdown;
+                }
             };
+        }
+    );
+
+    // @ts-ignore
+    const $beautify = asAsync_628dffd9c1e74e5cb82620a2c575e5dd(
+        async () => {
+            const EDITOR = $vs.window.activeTextEditor;
+            if (EDITOR) {
+                const DOC = EDITOR.document;
+                if (DOC) {
+                    let result: any = Symbol('LANGUAGE_NOT_SUPPORTED');
+                    const TEXT: string = DOC.getText();
+
+                    const LANG: string = DOC.languageId;
+                    switch ($h.normalizeString(DOC.languageId)) {
+                        case 'css':
+                            {
+                                const beautify = require('js-beautify').css;
+
+                                result = beautify(TEXT);
+                            }
+                            break;
+
+                        case 'html':
+                            {
+                                const beautify = require('js-beautify').html;
+
+                                result = beautify(TEXT);
+                            }
+                            break;
+
+                        case 'javascript':
+                            {
+                                const beautify = require('js-beautify').js;
+
+                                result = beautify(TEXT);
+                            }
+                            break;
+                    }
+
+                    if (_.isSymbol(result)) {
+                        throw new Error(`Language '${ LANG }' is not supported!`);
+                    }
+
+                    return {
+                        '__neweditor_tm_19790905': Symbol('NEW_EDITOR'),
+                        'column': $vs.ViewColumn.Two,
+                        'lang': LANG,
+                        'text': $h.toStringSafe(result),
+                    };
+                }
+            }
+
+            throw new Error('No active editor found!');
         }
     );
 
@@ -565,6 +622,7 @@ async function showHelp_579c52a1992b472183db2fff8c764504() {
         md += '---- | ----------- | -------\n';
         md += '`$alert(msg)` | Shows a (wanring) popup. | `$alert("Hello, TM!")`\n';
         md += '`$asc(str)` | Handles a value as string and returns the ASCII (codes). | `$asc("T")`\n';
+        md += '`$beautify` | Beautifies the code in the active editor and opens the result in a new one. | `$beautify`\n';
         md += '`$cmd(id, ...args)` | Executes a [Visual Studio Code command](https://code.visualstudio.com/api/references/commands). | `$cmd("vscode.openFolder")`\n';
         md += '`$DELETE(url, body?, headers?)` | Starts a HTTP DELETE request. | `$DELETE("https://example.com/users/19861222")`\n';
         md += '`$emojis(search?)` | Returns a list of [emojis](https://www.npmjs.com/package/node-emoji), by using an optional filter. | `$emojis("heart")`\n';