+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (vscode-powertools)
 
+## 0.15.0 (January 30th, 2019; code execution)
+
+* added following functions for [Code Execution](https://github.com/egodigital/vscode-powertools/wiki/Tools#code-execution):
+  * `$beautify`: beautifies the code in the active editor and opens the result in a new one
+
 ## 0.14.0 (January 28th, 2019; scripts and code execution)
 
 * [scripts](https://github.com/egodigital/vscode-powertools/wiki/Scripts) can run via [CoffeeScript language](https://coffeescript.org) now